update: 推論コマンド実行を複合動作クラスの判定動作に統合
