+++ modules/motions/BackgroundPlaCameraAction.h
@@ -8,13 +8,13 @@
 #define BACKGROUND_PLACAMERA_ACTION_H
 
 #include "PlaCameraAction.h"
-#include "BackgroundDirectionDetector.h"
 #include "AngleRotation.h"
 #include "Robot.h"
 #include "FrameSave.h"
 #include "ImageUploader.h"
 #include <thread>
 #include <iostream>
+#include <nlohmann/json.hpp>
 #include "CompositeMotion.h"
 
 class BackgroundPlaCameraAction : public CompositeMotion {
@@ -50,8 +50,13 @@ class BackgroundPlaCameraAction : public CompositeMotion {
   int position = 0;                    // 撮影位置（0:正面, 1:右, 2:後ろ, 3:左）
   cv::Rect roi;                        // 動体検出用の注目領域
   const std::string detectionTargetPath
-      = "etrobocon2025/datafiles/detection_target";  // 判定用画像ディレクトリのパス
+      = "etrobocon2025/datafiles/detection_target";      // 判定用画像ディレクトリのパス
   const std::string detectionTargetName = "background";  // 風景向き判定用画像ファイル名
+  const std::string command
+      = "make -C etrobocon2025/yolo_detection_system/ background-detection";  // 風景推論用コマンド
+  const std::string resultFilePath
+      = "etrobocon2025/yolo_detection_system/result/"
+        "background_result.json";  // 風景の向きを判定した結果が出力されるファイルパス
 
   /**
    * @brief 前提条件を満たしているかチェックする
